% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{pv_db_disconnect}
\alias{pv_db_disconnect}
\title{Disconnect from a database}
\usage{
pv_db_disconnect(con)
}
\arguments{
\item{con}{A connection from \code{\link[=pv_db_connect]{pv_db_connect()}}.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Disconnect from a database
}
