% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{pv_db_tables}
\alias{pv_db_tables}
\title{List tables in a database}
\usage{
pv_db_tables(con)
}
\arguments{
\item{con}{A connection from \code{\link[=pv_db_connect]{pv_db_connect()}}.}
}
\value{
Character vector of table names.
}
\description{
List tables in a database
}
