% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pv_network}
\alias{pv_network}
\title{Simulated collaboration network}
\format{
A list of two data frames: \code{nodes} (\code{id}, \code{group}) and \code{links}
(\code{source}, \code{target}, \code{value}).
}
\usage{
pv_network
}
\description{
A 16-person, 4-team network for \code{\link[=pv_force]{pv_force()}}: people collaborate mostly
within their team, occasionally across.
}
\keyword{datasets}
