% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{pvchart-shiny}
\alias{pvchart-shiny}
\alias{pvchartOutput}
\alias{renderPvchart}
\title{Shiny bindings for polyviz charts}
\usage{
pvchartOutput(outputId, width = "100\%", height = "420px")

renderPvchart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{CSS sizes for the widget.}

\item{expr}{An expression returning a polyviz widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression?}
}
\value{
\code{pvchartOutput} returns a Shiny output element;
\code{renderPvchart} returns a Shiny render function.
}
\description{
Output and render functions for using polyviz D3 widgets within Shiny
apps.
}
